requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==2.1.1
frozenlist==1.3.3
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
idna==3.4
linkedin-scraper==2.9.1
lxml==4.9.2
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slackclient==2.9.4
tqdm==4.65.0
urllib3==1.26.16
yarl==1.9.2
